Added support for Server instances
There was no support for epics and their children for cloud instances. The cloud uses parent and child relationship while in DC instance we have the old epic and epic child relationship. Epics should now be supported for Cloud instances and DC instances versions 7.0+ (according to rovo) but I tested on 9.11 in instEnv.