Established Parent and Child relationship for Cloud and Server (Server only has one parent-child relation)
Restored functionality to Serevr for subtasks. Epics to Standard Issues do not have standard parent and child relationships because DC has a different term/schema for epic and epic children. Epic children refer to their epics as 'Epic Link' rather than 'Parent' as subtasks refer to standard issues in Jira. This means that this task is done if we are going all issues that we can assign to as parents. To have standard issues connect to Epics, we would need to expand the task to include all parent-child as well as epic-epic children. Also with our current apis, a standard issue in Jira DC has no way to know if it has any connected epic links because it does not show under fields or field values.